@@ -1,76 +1,77 @@
 <?php
 
-use Illuminate\Http\Request;
+use App\Http\Controllers\AccountController;
+use App\Http\Controllers\AuthController;
+use App\Http\Controllers\FileController;
+use App\Http\Controllers\MapController;
+use App\Http\Controllers\ParticipantController;
+use App\Http\Controllers\RideController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| is assigned the "api" middleware group. Enjoy building your API!
-|
-*/
-
-Route::post('login', 'Auth\AuthController@login');
+Route::post('login', [AuthController::class, 'login']);
+Route::post('register', [AuthController::class, 'register']);
 
 Route::middleware('auth:sanctum')->group(function () {
-    Route::group(['prefix' => 'status', 'as' => 'status.'], function () {
+    Route::group(['prefix' => 'user', 'as' => 'user.'], function () {
         // 유저 상태
-        Route::get('user', 'StatusController@user')->name('user');
+        Route::get('/', [AuthController::class, 'index']);
+        // 로그아웃
+        Route::post('logout', [AuthController::class, 'logout']);
 
         // 라이드 참가 상태
-        Route::get('participation', 'StatusController@participation')->name('participation');
+        Route::get('participation', [UserController::class, 'participation']);
     });
 
     // 마이 페이지
     Route::group(['prefix' => 'account', 'as' => 'account.'], function () {
         // 회원 정보 업데이트
-        Route::put('/', 'AccountController@update')->name('update');
+        Route::put('/', [AccountController::class, 'update']);
         // 회원 탈퇴
-        Route::delete('/', 'AccountController@destroy')->name('destroy');
+        Route::delete('/', [AccountController::class, 'destroy']);
 
         // 개설 내역
-        Route::get('manage', 'AccountController@manage')->name('manage');
+        Route::get('manage', [AccountController::class, 'manage']);
         // 참가 내역
-        Route::get('attend', 'AccountController@attend')->name('attend');
+        Route::get('attend', [AccountController::class, 'attend']);
     });
 
     // 업로드
-    Route::group(['prefix' => 'upload', 'as' => 'upload.'], function () {
-        // GPX 파일 업로드
-        Route::post('gpx', 'FileController@gpx')->name('gpx');
+    Route::middleware('throttle:5,1')->group(function () {
+        Route::group(['prefix' => 'upload', 'as' => 'upload.'], function () {
+            // GPX 파일 업로드
+            Route::post('gpx', [FileController::class, 'gpx']);
 
-        // 썸네일 업로드
-        Route::post('thumbnail', 'FileController@thumbnail')->name('thumbnail');
+            // 썸네일 업로드
+            Route::post('thumbnail', [FileController::class, 'thumbnail']);
+        });
     });
 
     // 네이버 지도 리버스 지오코드
-    Route::get('reverse-geocode', 'MapController@reverse_geocode');
+    Route::get('reverse-geocode', [MapController::class, 'reverse_geocode']);
 });
+
 
 Route::group(['prefix' => 'ride', 'as' => 'ride.'], function () {
     // 라이드 리스트
-    Route::get('/', 'RideController@index')->name('index');
+    Route::get('/', [RideController::class, 'index']);
 
     // 라이드 상세
-    Route::get('{ride}', 'RideController@show')->name('show');
+    Route::get('{ride}', [RideController::class, 'show']);
 
     Route::middleware('auth:sanctum')->group(function () {
         // 라이드 저장
-        Route::post('store', 'RideController@store')->name('store');
+        Route::post('/', [RideController::class, 'store']);
         // 라이드 수정
-        Route::get('edit/{ride}', 'RideController@edit')->name('edit');
+        Route::get('{ride}/edit', [RideController::class, 'edit']);
         // 라이드 업데이트
-        Route::put('{ride}', 'RideController@update')->name('update');
+        Route::put('{ride}', [RideController::class, 'update']);
         // 라이드 삭제
-        Route::delete('{ride}', 'RideController@destroy')->name('destroy');
+        Route::delete('{ride}', [RideController::class, 'destroy']);
 
         // 라이드 참가
-        Route::post('attend', 'ParticipantController@store')->name('attend');
+        Route::post('attend', [ParticipantController::class, 'store']);
         // 라이드 참가 취소
-        Route::post('cancel', 'ParticipantController@destroy')->name('cancel');
+        Route::post('cancel', [ParticipantController::class, 'destroy']);
     });
 });
